r/key_vault: switching to use a state upgrader
